refactor: move bit related functions to core.reachable
