+++ solver/numlin.py
@@ -1,12 +1,12 @@
 """The Numberlink solver."""
 
-from math import log2
 from typing import Dict, List, Tuple, Union
 
 from .core.common import direction, display, fill_path, grid, shade_c
 from .core.penpa import Puzzle, Solution
 from .core.loop import single_loop
 from .core.neighbor import adjacent
+from .core.reachable import clue_bit, num_binary_range, grid_bit_color_connected
 from .core.helper import tag_encode
 from .core.solution import solver
 
@@ -28,36 +28,6 @@ def no_2x2_path_bit() -> str:
     rule += "no_empty(R, C) :- grid(R, C), bit_range(B), not bit_no(R, C, B).\n"
     rule += ":- grid(R, C), no_empty(R, C), not no_2x2(R, C).\n"
     return rule.strip()
-
-
-def clue_bit(r: int, c: int, _id: int, nbit: int) -> str:
-    """Assign clues with bit ids instead of numerical ids."""
-    rule = f"clue({r}, {c}).\n"
-    for i in range(nbit):
-        if _id >> i & 1:
-            rule += f"clue_bit({r}, {c}, {i}).\n"
-    return rule.strip()
-
-
-def grid_bit_color_connected(nbit: int, color: str = "grid", adj_type: Union[int, str] = "loop") -> str:
-    """Generate a constraint to check the reachability of {color} cells starting from a source (bit version)."""
-    if type(adj_type) is int:
-        adj_type = str(adj_type)
-    tag = tag_encode("reachable", "grid", "bit", "adj", adj_type)
-    rule = f"{{ {tag}(R, C, B) }} :- grid(R, C), {color}(R, C), bit_range(B).\n"
-    rule += f"{tag}(R, C, B) :- clue_bit(R, C, B).\n"
-    rule += f"not {tag}(R, C, B) :- grid(R, C), {color}(R, C), bit_range(B), clue(R, C), not clue_bit(R, C, B).\n"
-    rule += f"{tag}(R, C, B) :- {tag}(R1, C1, B), grid(R, C), bit_range(B), {color}(R, C), adj_{adj_type}(R, C, R1, C1).\n"
-    rule += f"not {tag}(R, C, B) :- not {tag}(R1, C1, B), grid(R, C), grid(R1, C1), bit_range(B), {color}(R, C), {color}(R1, C1), adj_{adj_type}(R, C, R1, C1).\n"
-    return rule.strip()
-
-
-def num_binary_range(num):
-    """Generate a rule restricting number represented by bits between 0 and num."""
-    nbit = int(log2(num)) + 1
-    rule = f"bit_range(0..{nbit - 1}).\n"
-    rule += "binary(0..1).\n"
-    return rule.strip(), nbit
 
 
 def solve(puzzle: Puzzle) -> List[Solution]:
